refactor: convert `pipeline.rs` from transport protocol to directory module
