+++ transport/hopr/benches/protocol_throughput_emulated_bench.rs
@@ -129,15 +129,13 @@ pub fn protocol_throughput_sender(c: &mut Criterion) {
                             codec_config,
                         );
 
-                        let processes = hopr_transport::protocol::PacketPipelineBuilder::new(
-                            PEERS[TESTED_PEER_ID].clone(),
-                            (wire_out_tx, wire_in_rx),
-                            (encoder, decoder),
-                            (api_recv_tx, api_send_rx),
-                            Default::default(),
-                        )
-                        .with_ticket_processing(ticket_proc, received_ack_tickets_tx)
-                        .build_for_relay();
+                        let processes =
+                            hopr_transport::protocol::PacketPipelineBuilder::new(PEERS[TESTED_PEER_ID].clone())
+                                .transport((wire_out_tx, wire_in_rx))
+                                .codec((encoder, decoder))
+                                .api((api_recv_tx, api_send_rx))
+                                .with_ticket_processing(ticket_proc, received_ack_tickets_tx)
+                                .build_for_relay();
 
                         let path = resolve_mock_path(
                             PEERS_CHAIN[TESTED_PEER_ID].public().to_address(),
